+++ data_layer/schema.js
@@ -33,7 +33,6 @@ async function createSchema(schema_create_object) {
 
         return schema_structure;
     } catch(err) {
-        logger.error(err);
         throw err;
     }
 }
@@ -44,7 +43,7 @@ async function createSchemaStructure(schema_create_object) {
         throw validation_error;
     }
 
-    if (global.hdb_schema[schema_create_object.schema]) {
+    if (!hdb_util.checkSchemaExists(schema_create_object.schema)) {
         throw `schema ${schema_create_object.schema} already exists`;
     }
 
@@ -66,7 +65,6 @@ async function createTable(create_table_object) {
 
         return create_table_structure;
     } catch(err) {
-        logger.error(err);
         throw err;
     }
 }
@@ -80,11 +78,12 @@ async function createTableStructure(create_table_object) {
 
     validation.validateTableResidence(create_table_object.residence);
 
-    if (!global.hdb_schema[create_table_object.schema]) {
-        throw `schema ${create_table_object.schema} does not exist`;
+    let check_schema_exists = hdb_util.checkSchemaExists(create_table_object.schema);
+    if (check_schema_exists) {
+        throw check_schema_exists;
     }
 
-    if (global.hdb_schema[create_table_object.schema][create_table_object.table]) {
+    if (!hdb_util.checkTableExists(create_table_object.schema, create_table_object.table)) {
         throw `table ${create_table_object.table} already exists in schema ${create_table_object.schema}`;
     }
 
@@ -119,6 +118,11 @@ async function dropSchema(drop_schema_object) {
         throw validation_error;
     }
 
+    let check_schema_exists = hdb_util.checkSchemaExists(drop_schema_object.schema);
+    if (check_schema_exists) {
+        throw check_schema_exists;
+    }
+
     try {
         await harperBridge.dropSchema(drop_schema_object);
         signalling.signalSchemaChange(signalling.SCHEMA_CHANGE_MESSAGE);
@@ -126,8 +130,7 @@ async function dropSchema(drop_schema_object) {
         const SCHEMA_DELETE_MSG = `successfully deleted schema ${drop_schema_object.schema}`;
 
         return SCHEMA_DELETE_MSG;
-    } catch(err) {
-        logger.error(err);
+    } catch (err) {
         throw err;
     }
 }
@@ -138,13 +141,18 @@ async function dropTable(drop_table_object) {
         throw validation_error;
     }
 
+    let check_schema_table_exist = hdb_util.checkSchemaTableExist(drop_table_object.schema, drop_table_object.table);
+    if (check_schema_table_exist) {
+        throw check_schema_table_exist;
+    }
+
     try {
         await harperBridge.dropTable(drop_table_object);
         signalling.signalSchemaChange({type: 'schema'});
         const TABLE_DELETE_MSG = `successfully deleted table ${drop_table_object.schema}.${drop_table_object.table}`;
 
         return TABLE_DELETE_MSG;
-    } catch(err) {
+    } catch (err) {
         logger.error(err);
         throw err;
     }
